unfinished createChildImpl migration
